@@ -1,14 +1,13 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
         applicationId "com.salamander.app_libs"
         minSdkVersion 18
-        targetSdkVersion 28
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
@@ -28,5 +27,6 @@ dependencies {
     //implementation 'com.google.android.material:material:1.1.0-alpha06'
     //implementation 'com.salamander:location:0.0.1-beta01'
     //implementation 'com.salamander:location:0.0.1'
-    implementation project(':network')
+    implementation project(':location')
+    //implementation project(':logger')
 }
