@@ -1,13 +1,13 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "0.0.1"
+        minSdkVersion 18
+        targetSdkVersion 29
+        versionCode 21021613
+        versionName "0.0.3-beta01"
     }
 
     buildTypes {
@@ -32,7 +32,8 @@ repositories {
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     api 'androidx.legacy:legacy-support-v4:1.0.0'
-    api 'com.google.android.material:material:1.1.0-alpha06'
+    api 'com.google.android.material:material:1.3.0'
+    api 'com.jakewharton.threetenabp:threetenabp:1.2.1'
 }
 
 apply from: 'publish.gradle'
