@@ -1,13 +1,13 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
 
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 28
-        versionCode 1
-        versionName "0.0.1-beta01"
+        minSdkVersion 18
+        targetSdkVersion 29
+        versionCode 21021613
+        versionName "0.0.3-beta01"
         multiDexEnabled true
     }
 
@@ -28,8 +28,9 @@ android {
 
 dependencies {
     implementation fileTree(include: ['*.jar'], dir: 'libs')
-    api 'com.salamander:network:0.0.1-beta01'
-    api 'com.google.android.gms:play-services-location:17.0.0'
+    api project(path: ':network')
+    //api 'com.salamander:network:0.0.2-beta03'
+    api 'com.google.android.gms:play-services-location:17.1.0'
     api 'com.jaredrummler:android-processes:1.1.1'
     implementation 'io.nlopez.smartlocation:library:3.3.3'
 }
