@@ -14,6 +14,7 @@ import androidx.appcompat.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.widget.ProgressBar;
 import android.widget.TextView;
+import com.salamander.core.R;
 
 /*
  * Copyright 2018 Qi Li
@@ -93,13 +94,10 @@ public class SalamanderProgressDialog extends DialogFragment {
         mStopMillisecond = Long.MAX_VALUE;
 
         showHandler = new Handler();
-        showHandler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                // only show if not already cancelled
-                if (mStopMillisecond > System.currentTimeMillis())
-                    showDialogAfterDelay();
-            }
+        showHandler.postDelayed(() -> {
+            // only show if not already cancelled
+            if (mStopMillisecond > System.currentTimeMillis())
+                showDialogAfterDelay();
         }, DELAY_MILLISECOND);
     }
 
@@ -136,12 +134,7 @@ public class SalamanderProgressDialog extends DialogFragment {
     private void cancelWhenShowing() {
         if (mStopMillisecond < mStartMillisecond + DELAY_MILLISECOND + MINIMUM_SHOW_DURATION_MILLISECOND) {
             final Handler handler = new Handler();
-            handler.postDelayed(new Runnable() {
-                @Override
-                public void run() {
-                    dismiss();
-                }
-            }, MINIMUM_SHOW_DURATION_MILLISECOND);
+            handler.postDelayed(this::dismiss, MINIMUM_SHOW_DURATION_MILLISECOND);
         } else {
             dismiss();
         }
@@ -149,12 +142,7 @@ public class SalamanderProgressDialog extends DialogFragment {
 
     private void cancelWhenNotShowing() {
         final Handler handler = new Handler();
-        handler.postDelayed(new Runnable() {
-            @Override
-            public void run() {
-                dismiss();
-            }
-        }, DELAY_MILLISECOND);
+        handler.postDelayed(this::dismiss, DELAY_MILLISECOND);
     }
 
     @Override
