@@ -1,6 +1,7 @@
 package com.salamander.app_libs;
 
 import android.app.Activity;
+import android.content.Context;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
@@ -9,15 +10,18 @@ import android.widget.EditText;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.salamander.core.Utils;
-import com.salamander.core.utils.DialogUtils;
-import com.salamander.network.DownloadCertificate;
 import com.salamander.network.retro.Retro;
 import com.salamander.network.retro.RetroData;
 import com.salamander.network.retro.RetroResp;
 import com.salamander.network.retro.RetroStatus;
 
+import java.util.concurrent.TimeUnit;
+
+import okhttp3.OkHttpClient;
 import okhttp3.ResponseBody;
+import okhttp3.logging.HttpLoggingInterceptor;
 import retrofit2.Call;
+import retrofit2.Retrofit;
 import retrofit2.http.GET;
 
 public class MainActivity extends AppCompatActivity {
@@ -38,16 +42,6 @@ public class MainActivity extends AppCompatActivity {
         final EditText tx_test2 = findViewById(R.id.tx_test2);
 
         tx_test.setText("0");
-        tx_test2.setText("0");
-
-        new DownloadCertificate(this, new DownloadCertificate.PostDownload() {
-            @Override
-            public void downloadDone(String errorMessage) {
-                if (errorMessage != null) {
-                    DialogUtils.showDialog(activity, errorMessage, true);
-                }
-            }
-        }).execute();
 
         bt_test.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -66,22 +60,35 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private interface getData {
-        @GET("check_version")
+        @GET("Packing_List")
         Call<ResponseBody> get();
     }
 
     private void getData(final Retro.OnCB onCB) {
-        String url = "https://intraweb.datascrip.co.id/go-sam/android_data/";
-        //url = "http://dtswebapi.datascrip.co.id/demo/rest_api/application/models/";
-        getData getData = Retro.createRetrofit(this, url).create(MainActivity.getData.class);
+        String url = "http://dtsnavtest2015.datascrip.co.id:8047/servertest_trace/WS/PT%20DATASCRIP/Page/";
+        getData getData = createRetrofit(this, url).create(MainActivity.getData.class);
         getData.get().enqueue(new RetroResp.SuccessCallback<ResponseBody>(activity) {
             @Override
             public void onCall(RetroData retroData) {
                 super.onCall(retroData);
                 if (Utils.isEmpty(retroData.getRetroStatus().getMessage()))
                     retroData.getRetroStatus().setMessage(retroData.getResult());
-                onCB.onCB(retroData.getRetroStatus());
+                //onCB.onCB(retroData.getRetroStatus());
             }
         });
+    }
+
+    private Retrofit createRetrofit(Context context, String URL) {
+        HttpLoggingInterceptor interceptor = new HttpLoggingInterceptor();
+        interceptor.setLevel(HttpLoggingInterceptor.Level.BODY);
+        OkHttpClient.Builder client = new OkHttpClient.Builder()
+                .authenticator(new NTLMAuthenticator("dbadmin", "K@m!sDinIHari7M@r2019"))
+                .addInterceptor(interceptor)
+                .connectTimeout(30, TimeUnit.SECONDS)
+                .readTimeout(2, TimeUnit.MINUTES);
+        return new Retrofit.Builder()
+                .baseUrl(URL)
+                .client(client.build())
+                .build();
     }
 }
